+++ field_history/managers.py
@@ -10,7 +10,4 @@ class FieldHistoryManager(Manager):
                            content_type=content_type)
 
     def get_for_model_and_field(self, object, field):
-        content_type = ContentType.objects.get_for_model(object)
-        return self.filter(object_id=object.pk,
-                           content_type=content_type,
-                           field_name=field)
+        return self.get_for_model(object).filter(field_name=field)